app/backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
# Day 13 Observability
structlog>=24.0.0
langfuse>=2.0.0
httpx>=0.25.0
